fix: [#27] UserEntity 수정
- ERD 설계한거에 맞춰서 Entity 수정했습니다
