+++ backend/src/main/java/org/example/config/WebOAuthSecurityConfig.java
@@ -4,8 +4,8 @@ import org.example.config.jwt.TokenProvider;
 import org.example.config.oauth.OAuth2AuthorizationRequestBasedOnCookieRepository;
 import org.example.config.oauth.OAuth2SuccessHandler;
 import org.example.config.oauth.OAuth2UserCustomService;
-import org.example.user.application.member.UserService;
 import org.example.user.repository.token.RefreshTokenRepository;
+import org.example.user.service.member.UserService;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.http.HttpStatus;
@@ -13,12 +13,11 @@ import org.springframework.security.config.annotation.web.builders.HttpSecurity;
 import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
 import org.springframework.security.config.annotation.web.configuration.WebSecurityCustomizer;
 import org.springframework.security.config.http.SessionCreationPolicy;
-import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
-import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.security.web.SecurityFilterChain;
 import org.springframework.security.web.authentication.HttpStatusEntryPoint;
 import org.springframework.security.web.authentication.UsernamePasswordAuthenticationFilter;
 import org.springframework.security.web.util.matcher.AntPathRequestMatcher;
+
 import lombok.RequiredArgsConstructor;
 
 @RequiredArgsConstructor
@@ -26,77 +25,77 @@ import lombok.RequiredArgsConstructor;
 @EnableWebSecurity
 public class WebOAuthSecurityConfig {
 
-    private final OAuth2UserCustomService oAuth2UserCustomService;
-    private final TokenProvider tokenProvider;
-    private final RefreshTokenRepository refreshTokenRepository;
-    private final UserService userService;
+	private final OAuth2UserCustomService oAuth2UserCustomService;
+	private final TokenProvider tokenProvider;
+	private final RefreshTokenRepository refreshTokenRepository;
+	private final UserService userService;
 
-    @Bean
-    public WebSecurityCustomizer webSecurityCustomizer() {
-        // 스프링 시큐리티에서 정적 자원에 대한 접근을 허용
-        return (web) -> web.ignoring().requestMatchers("/img/**", "/css/**", "/js/**");
-    }
+	@Bean
+	public WebSecurityCustomizer webSecurityCustomizer() {
+		// 스프링 시큐리티에서 정적 자원에 대한 접근을 허용
+		return (web) -> web.ignoring().requestMatchers("/img/**", "/css/**", "/js/**");
+	}
 
-    @Bean
-    public SecurityFilterChain securityFilterChain(HttpSecurity http) throws Exception {
-        // 토큰 방식으로 인증을 하기 때문에 기존에 사용하던 폼로그인, 세션 비활성화
-        http.csrf(csrf -> csrf.disable())
-            .httpBasic(httpBasic -> httpBasic.disable())
-            .formLogin(formLogin -> formLogin.disable())
-            .logout(logout -> logout.disable());
+	@Bean
+	public SecurityFilterChain securityFilterChain(HttpSecurity http) throws Exception {
+		// 토큰 방식으로 인증을 하기 때문에 기존에 사용하던 폼로그인, 세션 비활성화
+		http.csrf(csrf -> csrf.disable())
+			.httpBasic(httpBasic -> httpBasic.disable())
+			.formLogin(formLogin -> formLogin.disable())
+			.logout(logout -> logout.disable());
 
-        http.sessionManagement(session -> session.sessionCreationPolicy(SessionCreationPolicy.STATELESS));
+		http.sessionManagement(session -> session.sessionCreationPolicy(SessionCreationPolicy.STATELESS));
 
-        //헤더를 확인할 커스텀 필터 추가
-        http.addFilterBefore(tokenAuthenticationFilter(), UsernamePasswordAuthenticationFilter.class);
+		//헤더를 확인할 커스텀 필터 추가
+		http.addFilterBefore(tokenAuthenticationFilter(), UsernamePasswordAuthenticationFilter.class);
 
-        // 토큰 재발급 URL은 인증 없이 접근 가능하도록 설정. 나머지 API URL은 인증 필요
-        http.authorizeHttpRequests(authorize -> authorize
-            .requestMatchers("/api/token").permitAll()
-            .requestMatchers("/api/**").authenticated()
-            .anyRequest().permitAll()
-        );
+		// 토큰 재발급 URL은 인증 없이 접근 가능하도록 설정. 나머지 API URL은 인증 필요
+		http.authorizeHttpRequests(authorize -> authorize
+			.requestMatchers("/api/token").permitAll()
+			.requestMatchers("/api/**").authenticated()
+			.anyRequest().permitAll()
+		);
 
-        http.oauth2Login(oauth2 -> oauth2
-            .loginPage("/login")
-            .authorizationEndpoint(authorization -> authorization
-                .authorizationRequestRepository(oAuth2AuthorizationRequestBasedOnCookieRepository())
-            )
-            .successHandler(oAuth2SuccessHandler())
-            .userInfoEndpoint(userInfo -> userInfo
-                .userService(oAuth2UserCustomService)
-            )
-        );
+		http.oauth2Login(oauth2 -> oauth2
+			.loginPage("/login")
+			.authorizationEndpoint(authorization -> authorization
+				.authorizationRequestRepository(oAuth2AuthorizationRequestBasedOnCookieRepository())
+			)
+			.successHandler(oAuth2SuccessHandler())
+			.userInfoEndpoint(userInfo -> userInfo
+				.userService(oAuth2UserCustomService)
+			)
+		);
 
-        http.logout(logout -> logout
-            .logoutSuccessUrl("/login")
-        );
+		http.logout(logout -> logout
+			.logoutSuccessUrl("/login")
+		);
 
-        // /api로 시작하는 url인 경우 401 상태 코드를 반환하도록 예외 처리
-        http.exceptionHandling(exception -> exception
-            .defaultAuthenticationEntryPointFor(new HttpStatusEntryPoint(HttpStatus.UNAUTHORIZED),
-                new AntPathRequestMatcher("/api/**"))
-        );
+		// /api로 시작하는 url인 경우 401 상태 코드를 반환하도록 예외 처리
+		http.exceptionHandling(exception -> exception
+			.defaultAuthenticationEntryPointFor(new HttpStatusEntryPoint(HttpStatus.UNAUTHORIZED),
+				new AntPathRequestMatcher("/api/**"))
+		);
 
-        return http.build();
-    }
+		return http.build();
+	}
 
-    @Bean
-    public OAuth2SuccessHandler oAuth2SuccessHandler() {
-        return new OAuth2SuccessHandler(tokenProvider,
-            refreshTokenRepository,
-            oAuth2AuthorizationRequestBasedOnCookieRepository(),
-            userService
-        );
-    }
+	@Bean
+	public OAuth2SuccessHandler oAuth2SuccessHandler() {
+		return new OAuth2SuccessHandler(tokenProvider,
+			refreshTokenRepository,
+			oAuth2AuthorizationRequestBasedOnCookieRepository(),
+			userService
+		);
+	}
 
-    @Bean
-    public TokenAuthenticationFilter tokenAuthenticationFilter() {
-        return new TokenAuthenticationFilter(tokenProvider);
-    }
+	@Bean
+	public TokenAuthenticationFilter tokenAuthenticationFilter() {
+		return new TokenAuthenticationFilter(tokenProvider);
+	}
 
-    @Bean
-    public OAuth2AuthorizationRequestBasedOnCookieRepository oAuth2AuthorizationRequestBasedOnCookieRepository() {
-        return new OAuth2AuthorizationRequestBasedOnCookieRepository();
-    }
+	@Bean
+	public OAuth2AuthorizationRequestBasedOnCookieRepository oAuth2AuthorizationRequestBasedOnCookieRepository() {
+		return new OAuth2AuthorizationRequestBasedOnCookieRepository();
+	}
 }
