+++ backend/src/main/java/org/example/config/oauth/OAuth2SuccessHandler.java
@@ -4,10 +4,10 @@ import java.io.IOException;
 import java.time.Duration;
 
 import org.example.config.jwt.TokenProvider;
-import org.example.user.application.member.UserService;
 import org.example.user.domain.entity.member.UserEntity;
 import org.example.user.domain.entity.token.RefreshToken;
 import org.example.user.repository.token.RefreshTokenRepository;
+import org.example.user.service.member.UserService;
 import org.example.util.CookieUtil;
 import org.springframework.security.core.Authentication;
 import org.springframework.security.oauth2.core.user.OAuth2User;
@@ -23,56 +23,59 @@ import lombok.RequiredArgsConstructor;
 @Component
 public class OAuth2SuccessHandler extends SimpleUrlAuthenticationSuccessHandler {
 
-    public static final String REFRESH_TOKEN_COOKIE_NAME = "refresh_token";
-    public static final Duration REFRESH_TOKEN_DURATION = Duration.ofDays(14);
-    public static final Duration ACCESS_TOKEN_DURATION = Duration.ofDays(1);
-    public static final String REDIRECT_PATH = "/articles"; //redirect 경로
+	public static final String REFRESH_TOKEN_COOKIE_NAME = "refresh_token";
+	public static final Duration REFRESH_TOKEN_DURATION = Duration.ofDays(14);
+	public static final Duration ACCESS_TOKEN_DURATION = Duration.ofDays(1);
+	public static final String REDIRECT_PATH = "/articles"; //redirect 경로
 
-    private final TokenProvider tokenProvider;
-    private final RefreshTokenRepository refreshTokenRepository;
-    private final OAuth2AuthorizationRequestBasedOnCookieRepository authorizationRequestRepository;
-    private final UserService userService;
-    @Override
-    public void onAuthenticationSuccess(HttpServletRequest request, HttpServletResponse response, Authentication authentication) throws IOException {
-        OAuth2User oAuth2User = (OAuth2User) authentication.getPrincipal();
-        UserEntity user = userService.findByUsername((String) oAuth2User.getAttributes().get("name"));
+	private final TokenProvider tokenProvider;
+	private final RefreshTokenRepository refreshTokenRepository;
+	private final OAuth2AuthorizationRequestBasedOnCookieRepository authorizationRequestRepository;
+	private final UserService userService;
 
-        String refreshToken = tokenProvider.generateToken(user, REFRESH_TOKEN_DURATION);
-        saveRefreshToken(user.getId(), refreshToken);
-        addRefreshTokenToCookie(request, response, refreshToken);
+	@Override
+	public void onAuthenticationSuccess(HttpServletRequest request, HttpServletResponse response,
+		Authentication authentication) throws IOException {
+		OAuth2User oAuth2User = (OAuth2User)authentication.getPrincipal();
+		UserEntity user = userService.findByUsername((String)oAuth2User.getAttributes().get("name"));
 
-        String accessToken = tokenProvider.generateToken(user, ACCESS_TOKEN_DURATION);
-        String targetUrl = getTargetUrl(accessToken);
+		String refreshToken = tokenProvider.generateToken(user, REFRESH_TOKEN_DURATION);
+		saveRefreshToken(user.getUserId(), refreshToken);
+		addRefreshTokenToCookie(request, response, refreshToken);
 
-        clearAuthenticationAttributes(request, response);
+		String accessToken = tokenProvider.generateToken(user, ACCESS_TOKEN_DURATION);
+		String targetUrl = getTargetUrl(accessToken);
 
-        getRedirectStrategy().sendRedirect(request, response, targetUrl); //target url로 redirect
-    }
+		clearAuthenticationAttributes(request, response);
 
-    private void saveRefreshToken(Long userId, String newRefreshToken) {
-        RefreshToken refreshToken = refreshTokenRepository.findByUserId(userId)
-                .map(entity -> entity.update(newRefreshToken))
-                .orElse(new RefreshToken(userId, newRefreshToken));
+		getRedirectStrategy().sendRedirect(request, response, targetUrl); //target url로 redirect
+	}
 
-        refreshTokenRepository.save(refreshToken);
-    }
+	private void saveRefreshToken(Long userId, String newRefreshToken) {
+		RefreshToken refreshToken = refreshTokenRepository.findByUserId(userId)
+			.map(entity -> entity.update(newRefreshToken))
+			.orElse(new RefreshToken(userId, newRefreshToken));
 
-    private void addRefreshTokenToCookie(HttpServletRequest request, HttpServletResponse response, String refreshToken) {
-        int cookieMaxAge = (int) REFRESH_TOKEN_DURATION.toSeconds();
+		refreshTokenRepository.save(refreshToken);
+	}
 
-        CookieUtil.deleteCookie(request, response, REFRESH_TOKEN_COOKIE_NAME);
-        CookieUtil.addCookie(response, REFRESH_TOKEN_COOKIE_NAME, refreshToken, cookieMaxAge);
-    }
+	private void addRefreshTokenToCookie(HttpServletRequest request, HttpServletResponse response,
+		String refreshToken) {
+		int cookieMaxAge = (int)REFRESH_TOKEN_DURATION.toSeconds();
 
-    private void clearAuthenticationAttributes(HttpServletRequest request, HttpServletResponse response) {
-        super.clearAuthenticationAttributes(request);
-        authorizationRequestRepository.removeAuthorizationRequestCookies(request, response);
-    }
+		CookieUtil.deleteCookie(request, response, REFRESH_TOKEN_COOKIE_NAME);
+		CookieUtil.addCookie(response, REFRESH_TOKEN_COOKIE_NAME, refreshToken, cookieMaxAge);
+	}
 
-    private String getTargetUrl(String token) {
-        return UriComponentsBuilder.fromUriString(REDIRECT_PATH)
-                .queryParam("token", token)
-                .build()
-                .toUriString();
-    }
+	private void clearAuthenticationAttributes(HttpServletRequest request, HttpServletResponse response) {
+		super.clearAuthenticationAttributes(request);
+		authorizationRequestRepository.removeAuthorizationRequestCookies(request, response);
+	}
+
+	private String getTargetUrl(String token) {
+		return UriComponentsBuilder.fromUriString(REDIRECT_PATH)
+			.queryParam("token", token)
+			.build()
+			.toUriString();
+	}
 }
