+++ backend/src/main/java/org/example/post/domain/entity/PostEntity.java
@@ -1,9 +1,6 @@
 package org.example.post.domain.entity;
 
-import java.util.List;
-
 import org.example.common.TimeTrackableEntity;
-import org.example.post.domain.enums.PostStatus;
 import org.example.user.domain.entity.member.UserEntity;
 import org.hibernate.annotations.ColumnDefault;
 
@@ -15,17 +12,14 @@ import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
 import jakarta.persistence.JoinColumn;
 import jakarta.persistence.ManyToOne;
-import jakarta.persistence.OneToMany;
 import jakarta.persistence.Table;
 import lombok.AccessLevel;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-import lombok.Setter;
 
 @Entity
 @Table(name = "post")
 @Getter
-@Setter
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 public class PostEntity extends TimeTrackableEntity {
 
@@ -34,21 +28,21 @@ public class PostEntity extends TimeTrackableEntity {
 	@Column(name = "post_id")
 	private Long postId;
 
-	@ManyToOne(fetch = FetchType.EAGER)
-	@JoinColumn(name = "user_id", nullable = false)
+	@ManyToOne(fetch = FetchType.LAZY)
+	@JoinColumn(name = "user_id")
 	private UserEntity user;
 
-	@Column(name = "post_content", nullable = true, columnDefinition = "VARCHAR(255)")
+	@Column(name = "post_content", columnDefinition = "VARCHAR(255)")
 	private String postContent;
 
-	@Column(name = "post_status", nullable = false)
-	@ColumnDefault("0")
-	private PostStatus postStatus = PostStatus.PUBLISHED;
+	// @Column(name = "post_status")
+	// @ColumnDefault("0")
+	// private PostStatus postStatus = PostStatus.PUBLISHED;
 
-	@Column(name = "like_count", nullable = false, columnDefinition = "INT")
+	@Column(name = "like_count", columnDefinition = "INT")
 	@ColumnDefault("0")
 	private Integer likeCount = 0;
 
-	@OneToMany(mappedBy = "post", fetch = FetchType.LAZY)
-	private List<UserPostLikesEntity> likesList;
+	// @OneToMany(mappedBy = "post", fetch = FetchType.LAZY)
+	// private List<UserPostLikesEntity> likesList;
 }
