--- backend/src/main/java/org/example/user/application/token/RefreshTokenService.java
@@ -1,18 +1,0 @@
-package org.example.user.application.token;
-
-import org.example.user.domain.entity.token.RefreshToken;
-import org.example.user.repository.token.RefreshTokenRepository;
-import org.springframework.stereotype.Service;
-
-import lombok.RequiredArgsConstructor;
-
-@RequiredArgsConstructor
-@Service
-public class RefreshTokenService {
-    private final RefreshTokenRepository refreshTokenRepository;
-
-    public RefreshToken findByRefreshToken(String refreshToken) {
-        return refreshTokenRepository.findByRefreshToken(refreshToken).orElseThrow(() -> new IllegalArgumentException("unexpected token"));
-    }
-
-}
