+++ backend/src/main/java/org/example/user/controller/member/UserViewController.java
@@ -1,0 +1,23 @@
+package org.example.user.controller.member;
+
+import org.springframework.stereotype.Controller;
+import org.springframework.web.bind.annotation.GetMapping;
+
+@Controller
+public class UserViewController {
+
+	@GetMapping("/")
+	public String hello() {
+		return "index";
+	}
+
+	@GetMapping("/login")
+	public String login() {
+		return "oauthLogin";
+	}
+
+	@GetMapping("/signup")
+	public String signup() {
+		return "signup";
+	}
+}
