+++ backend/src/main/java/org/example/user/controller/token/TokenApiController.java
@@ -1,8 +1,8 @@
-package org.example.user.presentation.token;
+package org.example.user.controller.token;
 
-import org.example.user.application.token.TokenService;
 import org.example.user.domain.dto.request.token.CreateAccessTokenRequest;
 import org.example.user.domain.dto.response.token.CreateAccessTokenResponse;
+import org.example.user.service.token.TokenService;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.PostMapping;
@@ -14,13 +14,14 @@ import lombok.RequiredArgsConstructor;
 @RequiredArgsConstructor
 @RestController
 public class TokenApiController {
-    private final TokenService tokenService;
+	private final TokenService tokenService;
 
-    @PostMapping("/api/token")
-    public ResponseEntity<CreateAccessTokenResponse> createNewAccessToken(@RequestBody CreateAccessTokenRequest request) {
-        String newAccessToken = tokenService.createNewAccessToken(request.getRefreshToken());
+	@PostMapping("/api/token")
+	public ResponseEntity<CreateAccessTokenResponse> createNewAccessToken(
+		@RequestBody CreateAccessTokenRequest request) {
+		String newAccessToken = tokenService.createNewAccessToken(request.getRefreshToken());
 
-        return ResponseEntity.status(HttpStatus.CREATED)
-                .body(new CreateAccessTokenResponse(newAccessToken));
-    }
+		return ResponseEntity.status(HttpStatus.CREATED)
+			.body(new CreateAccessTokenResponse(newAccessToken));
+	}
 }
