+++ backend/src/main/java/org/example/user/domain/enums/UserStatus.java
@@ -1,13 +1,15 @@
 package org.example.user.domain.enums;
 
-import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
 @Getter
 @NoArgsConstructor
-@AllArgsConstructor
 public enum UserStatus {
 	USER_STATUS_SIGNED_IN("SIGNIN"), USER_STATUS_SIGNED_OUT("SIGNOUT");
 	private String userStatusType;
+
+	UserStatus(String userSign) {
+		this.userStatusType = userSign;
+	}
 }
