--- backend/src/main/java/org/example/user/presentation/member/UserApiController.java
@@ -1,42 +1,0 @@
-package org.example.user.presentation.member;
-
-import java.util.List;
-
-import org.example.user.application.member.UserService;
-import org.example.user.domain.dto.request.member.AddUserRequest;
-import org.springframework.security.core.context.SecurityContextHolder;
-import org.springframework.security.web.authentication.logout.SecurityContextLogoutHandler;
-import org.springframework.stereotype.Controller;
-import org.springframework.ui.Model;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PostMapping;
-
-import jakarta.servlet.http.HttpServletRequest;
-import jakarta.servlet.http.HttpServletResponse;
-import lombok.RequiredArgsConstructor;
-
-@RequiredArgsConstructor
-@Controller
-public class UserApiController {
-
-    private final UserService userService;
-
-    @PostMapping("/user")
-    public String signup(AddUserRequest addUserRequest) {
-        userService.saveUser(addUserRequest); // 회원가입 메서드 호출
-        return "redirect:/login"; // 회원 가입이 완료된 이후에 로그인 페이지로 이동
-    }
-
-    @GetMapping("/logout")
-    public String logout(HttpServletRequest request, HttpServletResponse response) {
-        new SecurityContextLogoutHandler().logout(request, response, SecurityContextHolder.getContext().getAuthentication());
-        return "redirect:/login";
-    }
-
-    @GetMapping("/articles")
-    public String getArticles(Model model) {
-
-        return "articleList";
-    }
-
-}
