+++ backend/src/main/java/org/example/user/service/member/UserDetailService.java
@@ -1,4 +1,4 @@
-package org.example.user.application.member;
+package org.example.user.service.member;
 
 import org.example.user.repository.member.UserRepository;
 import org.springframework.security.core.userdetails.UserDetails;
@@ -12,11 +12,12 @@ import lombok.RequiredArgsConstructor;
 // 스프링 시큐리티에서 사용자 정보를 가져오는 인터페이스
 public class UserDetailService implements UserDetailsService {
 
-    private final UserRepository userRepository;
-    // 사용자 이름 (email)으로 사용자의 정보를 가져오는 메서드
-    @Override
-    public UserDetails loadUserByUsername(String username) {
-        return userRepository.findByUsername(username)
-                .orElseThrow(() -> new IllegalArgumentException(username));
-    }
+	private final UserRepository userRepository;
+
+	// 사용자 이름 (email)으로 사용자의 정보를 가져오는 메서드
+	@Override
+	public UserDetails loadUserByUsername(String username) {
+		return userRepository.findByUsername(username)
+			.orElseThrow(() -> new IllegalArgumentException(username));
+	}
 }
