+++ backend/src/main/java/org/example/user/service/member/UserService.java
@@ -1,4 +1,4 @@
-package org.example.user.application.member;
+package org.example.user.service.member;
 
 import org.example.user.domain.dto.request.member.AddUserRequest;
 import org.example.user.domain.entity.member.UserEntity;
@@ -12,27 +12,27 @@ import lombok.RequiredArgsConstructor;
 @Service
 public class UserService {
 
-    private final UserRepository userRepository;
+	private final UserRepository userRepository;
 
-    public Long saveUser(AddUserRequest addUserRequest) {
-        BCryptPasswordEncoder encoder = new BCryptPasswordEncoder();
+	public Long saveUser(AddUserRequest addUserRequest) {
+		BCryptPasswordEncoder encoder = new BCryptPasswordEncoder();
 
-        return userRepository.save(UserEntity.builder()
-                .username(addUserRequest.getUsername())
-                .password(encoder.encode(addUserRequest.getPassword()))
-                .build()).getId();
-    }
+		return userRepository.save(UserEntity.builder()
+			.username(addUserRequest.getUsername())
+			.password(encoder.encode(addUserRequest.getPassword()))
+			.build()).getUserId();
+	}
 
-    public UserEntity findById(Long userId) {
-        return userRepository.findById(userId)
-                .orElseThrow(() -> new IllegalArgumentException("Unexpected user"));
-    }
+	public UserEntity findById(Long userId) {
+		return userRepository.findById(userId)
+			.orElseThrow(() -> new IllegalArgumentException("Unexpected user"));
+	}
 
-    public UserEntity findByUsername(String username) {
-        System.out.println("Searching for user: " + username);
-        return userRepository.findByUsername(username)
-                .orElseThrow(() -> new IllegalArgumentException("unexpected user"));
-    }
+	public UserEntity findByUsername(String username) {
+		System.out.println("Searching for user: " + username);
+		return userRepository.findByUsername(username)
+			.orElseThrow(() -> new IllegalArgumentException("unexpected user"));
+	}
 
 }
 //    // 이 메서드는 새 사용자를 생성하고 저장합니다. 실제 사용 시 비밀번호 등 추가 정보 처리가 필요할 수 있습니다.
